+++ src/jtree.js
@@ -13,6 +13,6 @@ jtree.TreeNode = TreeNode
 jtree.NonTerminalNode = GrammarBackedNonTerminalNode
 jtree.TerminalNode = GrammarBackedTerminalNode
 
-jtree.getVersion = () => "14.0.1"
+jtree.getVersion = () => "14.1.0"
 
 module.exports = jtree